add user id

+++ app/Http/Controllers/Admin/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Auth;
 
 use App\Post;
 use App\Category;
@@ -26,7 +27,9 @@ class PostController extends Controller
      */
     public function index()
     {
-        $posts = Post::all();
+        // $posts = Post::all();
+        $user = Auth::user();
+        $posts = $user->posts;
 
         return view('admin.posts.index', compact('posts'));
     }
@@ -62,6 +65,10 @@ class PostController extends Controller
         $newPost->slug = $this->getSlug($data['title']);
 
         $newPost->published = isset($data['published']); // true o false
+
+        // associo l'utente al post
+        $newPost->user_id = Auth::id(); // mi restituisce l'id dell'utente loggato
+
         $newPost->save();
 
         // se ci sono dei tags associati, li associo al post appena creato
@@ -80,6 +87,10 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        if($post->user_id !== Auth::id()) {
+            abort(403);
+        }
+
         return view('admin.posts.show', compact('post'));
     }
 
@@ -91,6 +102,10 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if($post->user_id !== Auth::id()) {
+            abort(403);
+        }
+
         $categories = Category::all();
         $tags = Tag::all();
 
@@ -110,6 +125,10 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if($post->user_id !== Auth::id()) {
+            abort(403);
+        }
+
         // validazione
         $request->validate($this->validation);
         // aggiornamento
@@ -139,6 +158,10 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if($post->user_id !== Auth::id()) {
+            abort(403);
+        }
+
         $post->delete();
 
         return redirect()->route('admin.posts.index');
